Fix rebar.config.script crashes
Handle errors of rebar_utils:sh/2 correctly and abort the
rebar.config.script in a clean way.

The config script just crashed if any of the called scripts was not
available or failed in some way. This is especially ugly if fuserl is
just a dependency in another repo as the crash will even interrupt a
'rebar clean'.